Move `api/auth.ts` to `api/auth/hooks.ts`
