+++ frontend/src/components/chats/list.tsx
@@ -7,7 +7,7 @@ import dayjs from "dayjs"
 import { useChat } from "@/api/chats/chats"
 import { useIsUserOnline, useUser } from "@/api/users/hooks"
 import { useMessage } from "@/api/chats/messages"
-import { useMe } from "@/api/auth"
+import { useMe } from "@/api/auth/hooks"
 import { MarkdownToPlain } from "@/components/markdown/render"
 import { ChatSectionCollapsedContext } from "@/components/layout"
 import { Message } from "@/ws"
